requirements.txt
Django==4.2.7
Pillow==10.1.0
requests==2.31.0
python-decouple==3.8
google-generativeai==0.3.2
PyMuPDF==1.26.6
celery==5.3.4
redis==5.0.1